+++ front/src/helpers/estatus.ts
@@ -1,4 +1,4 @@
 import { Role } from "@/types";
 
-export const roleGlobal: Role = Role.ADMIN
+export const roleGlobal: Role = Role.PROFESSIONAL
 export const isAuthenticatedGlobal:boolean = false;
